requirements.txt
ephem
geocoder
requests
RPi.GPIO
pyserial
smbus